Refactor angular apps
All controllers were extracted to their own file
each under `controllers` folder. Then each app must include
whatever controllers are needed. Tests now are focused
on controllers instead of whole angular apps

In order to get one angular app per page, thus having
controllers being shared among apps, it was needed to use
'import' and 'export', which is ES6 syntax, supported by
most browsers since 2015. That being said, modules can be
imported without having to add '<script>' tag all over the
templates.

Note: `package-lock.json` has lots of changes due to upgraded
version of `karma` from 3.0.0 to 3.1.0, which supports modules.
