+++ package.json
@@ -11,7 +11,7 @@
     "jasmine": "^3.2.0",
     "karma-jasmine": "^1.1.2",
     "karma-chrome-launcher": "^2.2.0",
-    "karma": "^3.0.0"
+    "karma": "^3.1.0"
   },
   "devDependencies": {},
   "scripts": {
