+++ squad/frontend/templates/squad/build.jinja2
@@ -197,7 +197,7 @@
 {% endblock %}
 
 {% block javascript %}
-<script type="text/javascript" src='{{static("squad/build.js")}}'></script>
+<script type="module" src='{{static("squad/build.js")}}'></script>
 <script type="text/javascript">
 csrf_token = '{{ csrf_token }}'
 </script>
