+++ squad/frontend/templates/squad/build_metadata.jinja2
@@ -10,7 +10,7 @@
 {% endblock %}
 
 {% block javascript %}
-<script type="text/javascript" src='{{static("squad/build.js")}}'></script>
+<script type="module" src='{{static("squad/build.js")}}'></script>
 <script type="text/javascript">
 csrf_token = '{{ csrf_token }}'
 </script>
