+++ squad/frontend/templates/squad/compare.jinja2
@@ -80,7 +80,7 @@
 
 {% block javascript %}
 <script type="text/javascript" src='{{static("select2.js/select2.min.js")}}'></script>
-<script type="text/javascript" src='{{static("squad/compare.js")}}'></script>
+<script type="module" src='{{static("squad/compare.js")}}'></script>
 {% endblock %}
 
 {% block styles %}
