+++ squad/frontend/templates/squad/metrics.jinja2
@@ -87,6 +87,6 @@ csrf_token = '{{ csrf_token }}'
 user = '{{request.user}}'
 </script>
 <script type="text/javascript" src="{{static("chartjs/Chart.bundle.js")}}"></script>
-<script type="text/javascript" src='{{static("squad/charts.js")}}'></script>
+<script type="module" src='{{static("squad/charts.js")}}'></script>
 <script type="text/javascript" src='{{static("chartjs/chartjs-plugin-annotation.min.js")}}'></script>
 {% endblock %}
