+++ squad/frontend/templates/squad/testjobs.jinja2
@@ -71,7 +71,7 @@
 {% endblock %}
 
 {% block javascript %}
-<script type="text/javascript" src='{{static("squad/build.js")}}'></script>
+<script type="module" src='{{static("squad/build.js")}}'></script>
 <script type="text/javascript">
 csrf_token = '{{ csrf_token }}'
 </script>
