+++ squad/frontend/templates/squad/tests.jinja2
@@ -91,7 +91,7 @@
 {% endblock %}
 
 {% block javascript %}
-<script type="text/javascript" src='{{static("squad/build.js")}}'></script>
+<script type="module" src='{{static("squad/build.js")}}'></script>
 <script type="text/javascript" src='{{static("squad/table.js")}}'></script>
 <script type="text/javascript">
 csrf_token = '{{ csrf_token }}'
