+++ squad/frontend/templates/squad/thresholds.jinja2
@@ -16,7 +16,7 @@
 
 {% block javascript %}
 <script type="text/javascript" src='{{static("angularjs/angular-resource.min.js")}}'></script>
-<script type="text/javascript" src='{{static("squad/metric.threshold.js")}}'></script>
+<script type="module" src='{{static("squad/metric.threshold.js")}}'></script>
 <script type="text/javascript" src='{{static("select2.js/select2.min.js")}}'></script>
 <script type="text/javascript">
 csrf_token = '{{ csrf_token }}'
