+++ test/unit/test_charts.js
@@ -1,6 +1,31 @@
+import {ChartPanel as _ChartPanel, ChartsController} from '../../squad/frontend/static/squad/controllers/charts.js'
+
+var app = angular.module('chartsApp', []);
+
+app.config(['$locationProvider', function($locationProvider) {
+    $locationProvider.html5Mode({
+        enabled: true,
+        requireBase: false
+    })
+}])
+
+app.factory('ChartPanel', _ChartPanel);
+
+app.controller(
+    'ChartsController',
+    [
+        '$scope',
+        '$http',
+        '$location',
+        '$compile',
+        'ChartPanel',
+        ChartsController
+    ]
+);
+
 describe("ChartsController", function () {
 
-    beforeEach(module("SquadCharts"));
+    beforeEach(module("chartsApp"));
 
     var $controller, chartsController;
 
@@ -9,8 +34,8 @@ describe("ChartsController", function () {
     }));
 
     describe("ChartPanel", function () {
-
-        DATA = {}
+        var metric, data, environments, ChartPanel
+        var DATA = {}
         beforeEach(function() {
             metric = {"max":100,"name":":tests:","min":0,"label":"Test pass %"}
             data = {
@@ -34,57 +59,57 @@ describe("ChartsController", function () {
 
         it('tests if minDate works on empty environments', function () {
             // Constructor calls updateMinDate
-            chartPanel = new ChartPanel(metric, data, [])
+            var chartPanel = new ChartPanel(metric, data, [])
             expect(chartPanel.minDate).toEqual(Math.round(new Date() / 1000))
         });
 
         it('tests if minDate works on empty data', function () {
             data["hi6220-hikey"] = []
             // Constructor calls updateMinDate
-            chartPanel = new ChartPanel(metric, data, environments)
+            var chartPanel = new ChartPanel(metric, data, environments)
             expect(chartPanel.minDate).toEqual(Math.round(new Date() / 1000))
         });
 
         it('tests if minDate works', function () {
             // Constructor calls updateMinDate
-            chartPanel = new ChartPanel(metric, data, environments)
+            var chartPanel = new ChartPanel(metric, data, environments)
             expect(chartPanel.minDate).toEqual(1498242465)
         });
 
         it('tests if maxDate works on empty environments', function () {
             // Constructor calls updateMaxDate
-            chartPanel = new ChartPanel(metric, data, [])
+            var chartPanel = new ChartPanel(metric, data, [])
             expect(chartPanel.maxDate).toEqual(0)
         });
 
         it('tests if maxDate works on empty data', function () {
             data["hi6220-hikey"] = []
             // Constructor calls updateMaxDate
-            chartPanel = new ChartPanel(metric, data, environments)
+            var chartPanel = new ChartPanel(metric, data, environments)
             expect(chartPanel.maxDate).toEqual(0)
         });
 
         it('tests if maxDate works', function () {
             // Constructor calls updateMaxDate
-            chartPanel = new ChartPanel(metric, data, environments)
+            var chartPanel = new ChartPanel(metric, data, environments)
             expect(chartPanel.maxDate).toEqual(1498422468)
         });
 
         it('tests if filterData works on empty data', function () {
             data["hi6220-hikey"] = []
-            filtered_data = ChartPanel.filterData(
+            var filtered_data = ChartPanel.filterData(
                 data, 1498307267, 1498422458)
             expect(filtered_data.length).toEqual(0)
         });
 
         it('tests if filterData works on ambiguous limits', function () {
-            filtered_data = ChartPanel.filterData(
+            var filtered_data = ChartPanel.filterData(
                 data, 1498422458, 1498307267)
             expect(filtered_data.length).toEqual(0)
         });
 
         it('tests if filterData works', function () {
-            filtered_data = ChartPanel.filterData(
+            var filtered_data = ChartPanel.filterData(
                 data["hi6220-hikey"], 1498307267, 1498422458)
             expect(filtered_data.length).toEqual(4)
             expect(filtered_data).toEqual(
@@ -97,7 +122,7 @@ describe("ChartsController", function () {
         });
 
         it('tests if filterData works on ambiguous limits', function () {
-            filtered_data = ChartPanel.filterData(
+            var filtered_data = ChartPanel.filterData(
                 data, 1498422458, 1498307267)
             expect(filtered_data.length).toEqual(0)
         });
@@ -105,13 +130,12 @@ describe("ChartsController", function () {
     });
 
     describe('$scope.getEnvironmentIds', function() {
-        var $scope, controller, $attrs;
+        var $scope, controller, $attrs, $location;
 
         beforeEach(function() {
             $scope = {};
             $attrs = {};
             $location = "";
-            URL = {};
             controller = $controller('ChartsController', {
                 $scope: $scope,
                 $attrs: $attrs,
@@ -126,7 +150,7 @@ describe("ChartsController", function () {
                 {name: "juno-r2", line_color: "#563c66"},
                 {name: "x15", line_color: "#a40000", selected: true}
             ]
-            env_ids = $scope.getEnvironmentIds()
+            var env_ids = $scope.getEnvironmentIds()
             expect(env_ids.length).toEqual(3)
             expect($scope.getEnvironmentIds()).toEqual(
                 ["hi6220-hikey", "x86", "x15"])
@@ -135,13 +159,13 @@ describe("ChartsController", function () {
     });
 
     describe('$scope.download', function() {
-        var $scope, controller, $attrs, $httpBackend;
+        var $scope, controller, $attrs, $httpBackend, 
+            $location, responseData, obj
 
         beforeEach(function() {
             $scope = {};
             $attrs = {};
             $location = "";
-            URL = {};
             controller = $controller('ChartsController', {
                 $scope: $scope,
                 $attrs: $attrs,
